make log message arguments consistent to eliminate warnings

+++ trunk/src/slurmd/slurmstepd/req.c
@@ -1038,8 +1038,8 @@ _handle_attach(int fd, slurmd_job_t *job, uid_t uid)
 	srun       = xmalloc(sizeof(srun_info_t));
 	srun->key  = (srun_key_t *)xmalloc(SLURM_IO_KEY_SIZE);
 
-	debug("sizeof(srun_info_t) = %ld, sizeof(slurm_addr_t) = %ld",
-	      sizeof(srun_info_t), sizeof(slurm_addr_t));
+	debug("sizeof(srun_info_t) = %d, sizeof(slurm_addr_t) = %d",
+	      (int) sizeof(srun_info_t), (int) sizeof(slurm_addr_t));
 	safe_read(fd, &srun->ioaddr, sizeof(slurm_addr_t));
 	safe_read(fd, &srun->resp_addr, sizeof(slurm_addr_t));
 	safe_read(fd, srun->key, SLURM_IO_KEY_SIZE);
